Move to org.opentripplanner.routing.service
